refactor dump sql generation
